requirements.txt
streamlit
youtube-transcript-api
google-genai
ics